+++ src/vdom/attributes.js
@@ -1,10 +1,12 @@
 const {REGEX, ATTR, PROPS_ATTRIBUTES} = require('../constants');
 //const castStringTo = require('../utils/cast-string-to');
 const objectPath = require('../utils/object-path');
-const mapCompiled = require('./map-compiled');
+const isListener = require('../utils/is-listener');
+const mapper = require('./mapper');
+const {isDirective} = require('../directives/helpers');
 
 function isEventAttribute(name) {
-    return REGEX.IS_LISTENER.test(name);
+    return isListener(name);
 }
 
 function setAttribute($target, name, value, cmp) {
@@ -22,18 +24,20 @@ function setAttribute($target, name, value, cmp) {
         return;
     }
 
-    if (isCustomAttribute(name) || typeof value === 'function' || typeof value === 'object') {
+    if ((isCustomAttribute(name) || typeof value === 'function' || typeof value === 'object') && !isDirective(name)) {
+        // why? I need to remove any orphan keys in the mapper. Orphan keys are created by handler attributes
+        // like onclick, onmousedown etc. ...
+        // handlers are associated to the element only once.
+        // at the moment the only way to remove the keys is to take them.
+        if (isEventAttribute(name) && typeof value === 'string') {
+            mapper.getAll(value);
+        }
     } else if (typeof value === 'boolean') {
         setBooleanAttribute($target, name, value);
-    /*} else if (typeof value === 'object') {
-        try {
-            //$target.setAttribute(name, JSON.stringify(value));
-        } catch (e) {
-
-        }*/
     } else {
         if (value === undefined) value = '';
         $target.setAttribute(name, value);
+        //$target[name] = value;
     }
 }
 
@@ -76,7 +80,7 @@ function isCustomAttribute(name) {
 
 function setBooleanAttribute($target, name, value) {
     $target.setAttribute(name, value);
-    $target[name] = value;
+    //$target[name] = value;
 }
 
 function extractEventName(name) {
@@ -91,79 +95,96 @@ function addEventListener($target, name, value, cmp, cmpParent) {
 
     if (!isEventAttribute(name)) return;
 
-    // If use scope. from onDrawByParent event
-    let match = value.match(REGEX.GET_LISTENER_SCOPE);
+    let alreadyFunction = false;
 
-    if (match) {
+    // Determines if the function is passed by mapper
+    if (typeof value === 'function') {
+        alreadyFunction = true;
+    }
 
-        let args = null;
-        let handler = match[1];
-        let stringArgs = match[2];
-        if (stringArgs) {
-            args = stringArgs.split(',').map(item => {
-                item = trimQuotes(item.trim());
-                //return item === 'scope' ? cmpParent : castStringTo(trimQuotes(item))
-                let itemMap = mapCompiled.get(item);
-                if (itemMap !== undefined)
-                    item = itemMap;
-
-                return item === 'scope'
-                    ? cmpParent
-                    : item
-            })
-        }
-
-        const method = objectPath(handler, cmpParent);
-        if (method !== undefined) {
-            value = args
-                ? method.bind(cmpParent, ...args)
-                : method.bind(cmpParent);
-        }
-
-    } else {
-
-        match = value.match(REGEX.GET_LISTENER);
+    // Legacy logic where use a string instead of function
+    if (typeof value === 'string') {
+        // If use scope. from onDrawByParent event
+        let match = value.match(REGEX.GET_LISTENER_SCOPE);
 
         if (match) {
+
             let args = null;
             let handler = match[1];
             let stringArgs = match[2];
             if (stringArgs) {
                 args = stringArgs.split(',').map(item => {
                     item = trimQuotes(item.trim());
-                    let itemMap = mapCompiled.get(item);
+                    //return item === 'scope' ? cmpParent : castStringTo(trimQuotes(item))
+                    let itemMap = mapper.get(item);
                     if (itemMap !== undefined)
                         item = itemMap;
-                    //return item === 'this' ? cmp : castStringTo(trimQuotes(item))
-                    return item === 'this'
-                        ? cmp
+
+                    return item === 'scope'
+                        ? cmpParent
                         : item
                 })
             }
 
-            let isParentMethod = handler.match(REGEX.IS_PARENT_METHOD);
-
-            if (isParentMethod) {
-                handler = isParentMethod[1];
-                cmp = cmp.parent;
-            }
-
-            const method = objectPath(handler, cmp);
-
+            const method = objectPath(handler, cmpParent);
             if (method !== undefined) {
                 value = args
-                    ? method.bind(cmp, ...args)
-                    : method.bind(cmp);
+                    ? method.bind(cmpParent, ...args)
+                    : method.bind(cmpParent);
+            }
+
+        } else {
+
+            match = value.match(REGEX.GET_LISTENER);
+
+            if (match) {
+                let args = null;
+                let handler = match[1];
+                let stringArgs = match[2];
+                if (stringArgs) {
+                    args = stringArgs.split(',').map(item => {
+                        item = trimQuotes(item.trim());
+                        let itemMap = mapper.get(item);
+                        if (itemMap !== undefined)
+                            item = itemMap;
+                        //return item === 'this' ? cmp : castStringTo(trimQuotes(item))
+                        return item === 'this'
+                            ? cmp
+                            : item
+                    })
+                }
+
+                let isParentMethod = handler.match(REGEX.IS_PARENT_METHOD);
+
+                if (isParentMethod) {
+                    handler = isParentMethod[1];
+                    cmp = cmp.parent;
+                }
+
+                const method = objectPath(handler, cmp);
+
+                if (method !== undefined) {
+                    value = args
+                        ? method.bind(cmp, ...args)
+                        : method.bind(cmp);
+                }
             }
         }
     }
 
-    if (typeof value === 'function')
-        $target.addEventListener(
-            extractEventName(name),
-            value
-        );
-    else {
+    if (typeof value === 'function') {
+        if (alreadyFunction) {
+            $target.addEventListener(
+                extractEventName(name),
+                value.bind(cmp)
+            );
+        } else {
+            $target.addEventListener(
+                extractEventName(name),
+                value
+            );
+        }
+    } else {
         value = value.replace(REGEX.THIS_TARGET, '$target');
         // I don't understand but with regex test sometimes it don't works fine so use match... boh!
         //if (REGEX.IS_LISTENER_SCOPE.test(value) || value === 'scope') {
@@ -191,22 +212,21 @@ function addEventListener($target, name, value, cmp, cmpParent) {
 
 function attach($target, nodeProps, cmp, cmpParent) {
 
-    let bindValue;
     let name;
 
     const propsKeys = Object.keys(nodeProps);
 
     for(let i = 0, len = propsKeys.length; i < len; i++) {
         name = propsKeys[i];
-        setAttribute($target, name, nodeProps[name], cmp, cmpParent);
         addEventListener($target, name, nodeProps[name], cmp, cmpParent);
+        setAttribute($target, name, nodeProps[name], cmp, cmpParent);
         cmp.$$afterAttributeCreate($target, name, nodeProps[name], nodeProps);
     }
 
     const datasetArray = Object.keys($target.dataset);
     for (let i = 0; i < datasetArray.length; i++) {
-        if (REGEX.IS_LISTENER.test(datasetArray[i]))
-            addEventListener($target, i, $target.dataset[datasetArray[i]], cmp, cmpParent);
+        if (isListener(datasetArray[i]))
+            addEventListener($target, datasetArray[i], $target.dataset[datasetArray[i]], cmp, cmpParent);
     }
 
     //cmp.$$afterAttributesCreate($target, bindValue);