+++ src/vdom/mapper.js
@@ -1,10 +1,15 @@
+const RND = Math.random();
+
+const REGEX_1 = new RegExp('(\\/\\*' + RND + '=%{\\d+}%=\\*\\/)', 'g');
+const REGEX_2 = new RegExp('^\\/\\*' + RND + '=%{\\d+}%=\\*\\/$');
+
 module.exports = {
     lastId: 0,
     data: {},
-    set(value) {
+    set(value, from) {
         let id = ++this.lastId;
-        id = `/*=%{${id}}%=*/`;
-        //console.log('--->', value)
+        id = `/*${RND}=%{${id}}%=*/`;
+        //console.log('--->', id, value, from)
         this.data[id] = value;
         return id;
     },
@@ -13,10 +18,11 @@ module.exports = {
         id = id.trim();
         let res = this.data[id];
         delete this.data[id];
+        //this.flush()
         return res;
     },
     getAll(str) {
-        return str.replace(/(\/\*=%{\d+}%=\*\/)/g, (match) => {
+        return str.replace(REGEX_1, (match) => {
             let objValue = this.get(match);
             if (objValue !== undefined) {
                 return objValue;
@@ -25,7 +31,7 @@ module.exports = {
         });
     },
     isValidId(id) {
-        return /^\/\*=%{\d+}%=\*\/$/.test(id)
+        return REGEX_2.test(id)
     },
     flush() {
         this.data = {};